+++ includes/admin/views/view-settings-features.php
@@ -10,7 +10,7 @@ $listTable = new Features_List_Table([
         'description' => __('Description', 'oes'),
         'actions' => __('Actions', 'oes')
     ],
-    'meta_key' => 'field_oes_comment',
+    'meta_key' => 'title',
     'filter' => ['group', 'status']
 ]);